Moved confirm email fixtures to conftest
