+++ tests/functional/test_confirm_email_address.py
@@ -2,35 +2,6 @@ import pytest
 from flask import url_for
 
 
-@pytest.fixture
-def set_email_known(browser):
-    def do_set_email_known(known=True):
-        value = 'yes' if known else 'no'
-        browser.choose('email_known', value)
-    return do_set_email_known
-
-
-@pytest.fixture
-def submit_email_address(browser):
-    def do_submit_email_address(email_address):
-        browser.fill('email_address', email_address)
-        browser.find_by_css('form button').click()
-    return do_submit_email_address
-
-
-@pytest.fixture
-def submit_known_email_address(
-        browser,
-        set_email_known,
-        submit_email_address):
-
-    def do_submit_known_email_address(email_address):
-        browser.visit(url_for('main.request_email_address', _external=True))
-        set_email_known(True)
-        submit_email_address(email_address)
-    return do_submit_known_email_address
-
-
 @pytest.mark.usefixtures('live_server')
 class When_on_confirm_email_with_yes_selected_and_valid_email(object):
